Vytvořena funkce registrace, přihlášení a odhlášení, upravena možnost mazání a úpravy nabídky práce pouze zadavatelem

+++ base/views.py
@@ -1,7 +1,61 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, typprace
 from .forms import RoomForm
+
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('Domovskástránka')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Tento uživatel neexistuje!!')
+        
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('Domovskástránka')
+        else:
+            messages.error(request, "Uživatel neexistuje, nebo je špatné heslo!")
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('Domovskástránka')
+
+
+def registerUser(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('Domovskástránka')
+        else:
+            messages.error(request, "Nastala chyba při registraci, zkuste se registrovat znovu!")
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def Home(request):
@@ -74,7 +128,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'room.html', context)
 
-
+@login_required(login_url='login')
 def Vytvornabidku(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -86,10 +140,13 @@ def Vytvornabidku(request):
     context = {'form': form}
     return render(request, 'uprava_nabidky.html', context)
 
-
+@login_required(login_url='login')
 def Upraveninabidky(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.Zadávající:
+        return HttpResponse('Nemáte oprávnění upravovat cizí nabídky práce!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -99,9 +156,13 @@ def Upraveninabidky(request, pk):
     context = {'form': form}
     return render(request, 'uprava_nabidky.html', context)
 
-
+@login_required(login_url='login')
 def Smazaninabidky(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.Zadávající:
+        return HttpResponse('Nemáte oprávnění upravovat cizí nabídky práce!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('/Nabidky/0')
